[LE-15521] Missing <CR><LF> between the header and body if content-length is 0
Issue:
As per RFC 7230, an empty line must separate the headers and
optional body of an HTTP response.

While parsing the HTTP response, tinyHTTP uses this additional CRLF to
infer that all headers have been received. However, the additional CRLF
is then discarded and not forwarded to the command output.

This was counteracted within HLLEAPP-699 by adding an extra CRLF before
writing the HTTP response body to the serial port. However, in the case
where the content length is 0, the callback responsible for handling the
response body was never invoked. This resulted in a missing CRLF
following the headers.

Solution:
The extra CRLF is now appended following completion of the
header parsing rather than at the start of the body parsing.

Resolves: LE-15521
Related: EURY-3810 HLLEAPP-768
Change-Id: Iaa903a3d3a07edfa74163afd7b8d8308b9f7de81
Signed-off-by: Grayson Halliday <GHalliday@sierrawireless.com>