[LE-16160] Empty HTTP header handling for KHTTPGET
Issue:
An attempt to use +KHTTPGET to retrieve the google home page
(URL www.google.com)  from the home page (root path "/") fails
with +KHTTP_ERROR:1,5

Root Cause:
The HTTP parser in the tinyHttp 3rd party library does not
allow empty keyvalues in header key-value pairs. For example:
   Access-Control-Allow-Origin:  \r\n
   Access-Control-Allow-Credentials: \r\n
   Access-Control-Allow-Methods: \r\n
   Access-Control-Allow-Headers: \r\n
   Access-Control-Expose-Headers:  \r\n
When the parser state machine in http.c and header.c encounters
a CR or LF prematurely while in the state where it expects
a keyvalue, an error is returned and this results in
URC +KHTTP_ERROR:1,5

Solution:
The parser state machine for the header response should just
store the default key value 0 if a CR-LF is encountered
because of empty keyvalue after the header key instead of
returning an error.  In the http_header_state lookup table
in header.c, for state 6 (leading whitespace before header value),
change the opcode for input \n to 0xC4
(http_header_status_store_keyvalue) and change the opcode for
input \r to 10 (for state 10 CR after header value).

Resolves: LE-16160
Related: EURY-4366
Change-Id: Icadb6c5b805a6eebb8a34627187c349f2979a857
Signed-off-by: Gordon Finlay <gfinlay@sierrawireless.com